% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{generate_download_filename}
\alias{generate_download_filename}
\title{generate a filename for a download}
\usage{
generate_download_filename(template, root, data, create_directory = TRUE)
}
\arguments{
\item{template}{\link{whisker} template}

\item{root}{the root of the directory hierarchy}

\item{data}{data used to populate the template}

\item{create_directory}{boolean indicating whether to
(recursively) create the directory hierarchy.}
}
\value{
a filename
}
\description{
generate a filename for a download
}
