% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monsters.R
\docType{data}
\name{monsters}
\alias{monsters}
\title{Monsters}
\format{A tibble}
\usage{
data(monstr)
}
\description{
Monsters in the 5th edition SRD. Data accessible via the   
(\href{http://dnd.wizards.com/articles/features/systems-reference-document-srd}{Open Gaming Licence Version 1.0a}
}
\examples{
data(monsters)
head(monsters)
}
\keyword{datasets}
