% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{monstr_clean}
\alias{monstr_clean}
\title{Clean the data according to MONSTR rules.}
\usage{
monstr_clean(metadata)
}
\arguments{
\item{metadata}{description the downloaded file.}
}
\value{
description of the cleaned data
}
\description{
Clean the data according to MONSTR rules.
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
