% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{monstr_data}
\alias{monstr_data}
\title{Get the Data}
\usage{
monstr_data(metadata)
}
\arguments{
\item{metadata}{description of the downloaded data}
}
\value{
a \code{\link[tibble]{dplyr::tibble}} of the data from the
described download
}
\description{
Extract the tibble of the actual data
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
