% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{monstr_pipeline_defaults}
\alias{monstr_pipeline_defaults}
\title{Create the MONSTR defaults}
\usage{
monstr_pipeline_defaults(download_root = "")
}
\arguments{
\item{download_root}{Root of directory hierarchy.}
}
\value{
an augmented metadata
}
\description{
Create the MONSTR defaults
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
