% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{monstr_write_clean}
\alias{monstr_write_clean}
\title{Writes the data to the 'clean' area}
\usage{
monstr_write_clean(metadata, format = "csv", create_directory = TRUE)
}
\arguments{
\item{metadata}{description of the data.}

\item{format}{any known format or "all" to save a copy as all
known formats}

\item{create_directory}{boolean indicating whether directories
should be created.}
}
\value{
a boolean indicating success
}
\description{
Writes the data to the 'clean' area
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
