% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.R
\name{ons_api_call}
\alias{ons_api_call}
\title{Call the ONS API}
\usage{
ons_api_call(url)
}
\arguments{
\item{url}{url to call @seeAlso \code{\link{[api_base_url]}}}
}
\value{
a dataframe contained the API call results
}
\description{
Make request to given url, which is assumed to be the ONS api.
}
\details{
data retrieved is converted to tidyverse tibble if possible.
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
