% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.R
\name{ons_available_editions}
\alias{ons_available_editions}
\title{Available Editions}
\usage{
ons_available_editions(id)
}
\arguments{
\item{id}{dataset identifier. Valid values from \code{ons_available_datasets(...)}}
}
\value{
a list of edition identifiers
}
\description{
Available Editions
}
\examples{
\dontrun{
ons_available_editions(id = 'mid-year-pop-est')
}
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
