% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.R
\name{ons_datasets_setup}
\alias{ons_datasets_setup}
\title{Datasets Setup}
\usage{
ons_datasets_setup()
}
\value{
a dataframe describing available datasets
}
\description{
Retrieves a dataframe describing the datasets available from ONS via the API.
}
\details{
This returns a dataframe containing details that can be passed to other fns in this package for further processing
}
\author{
neale
}
