% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{safe_download}
\alias{safe_download}
\title{Safe Download

Downloads a file and tries hard to tidy up in the event of
errors. Since these files are typically large we don't want to
leave them in temp directories.

The destfile should only appear if the download was successful.}
\usage{
safe_download(url, destfile, fvalidate)
}
\arguments{
\item{url}{src for the download}

\item{destfile}{destination filename}

\item{fvalidate}{a fn that is passed the filename after download
to validate it in some way. The fn should return TRUE if the
file is valid.}
}
\description{
Safe Download

Downloads a file and tries hard to tidy up in the event of
errors. Since these files are typically large we don't want to
leave them in temp directories.

The destfile should only appear if the download was successful.
}
