% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{thf_clean}
\alias{thf_clean}
\title{Clean the data according to THF rules.}
\usage{
thf_clean(metadata)
}
\arguments{
\item{metadata}{description the downloaded file.}
}
\value{
description of the cleaned data
}
\description{
Clean the data according to THF rules.
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
