% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{thf_pipeline_defaults}
\alias{thf_pipeline_defaults}
\title{Apply THF defaults}
\usage{
thf_pipeline_defaults(df, download_root = "")
}
\arguments{
\item{df}{A 'base' setup, e.g. from \code{\link{ons_datasets_setup}}}

\item{download_root}{Root of directory hierarchy.}
}
\value{
an augmented dataframe
}
\description{
Apply THF defaults
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
