% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{thf_write_clean}
\alias{thf_write_clean}
\title{Writes the data to the 'clean' area}
\usage{
thf_write_clean(df, format = "csv", create_directory = TRUE)
}
\arguments{
\item{df}{dataframe describing the data.}

\item{format}{any known format or "all" to save a copy as all
known formats}

\item{create_directory}{boolean indicating whether directories
should be created.}
}
\value{
a boolean indicating success
}
\description{
Writes the data to the 'clean' area
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
