% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{write_csv}
\alias{write_csv}
\title{write the data as a csv.}
\usage{
write_csv(data, monstr, create_directory)
}
\arguments{
\item{data}{The actual data}

\item{monstr}{metadata dataframe created by the pipeline}

\item{create_directory}{boolean indicating whether to
(recursively) create the directory hierarchy.}
}
\value{
boolean indicating success
}
\description{
write the data as a csv.
}
\author{
Neale Swinnerton <neale@mastodonc.com
}
