% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.R
\name{ons_available_datasets}
\alias{ons_available_datasets}
\title{Available Datasets}
\usage{
ons_available_datasets(df)
}
\arguments{
\item{df}{dataframe describing the datasets from \code{ons_datasets_setup()}}
}
\value{

}
\description{
Available Datasets
}
\author{
neale
}
