% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.R
\name{ons_available_datasets}
\alias{ons_available_datasets}
\title{Available Datasets}
\usage{
ons_available_datasets()
}
\value{
list of available datasets and associated metadata
}
\description{
Retrieves a dataframe describing the datasets available from ONS via the API.
}
\examples{
\dontrun{
# return information on all available datasets and then filter on specific id
datasets <- ons_available_datasets()

datasets \%>\% 
filter(id='health-accounts')
}
\dontrun{
# display just the ids
ons_available_datasets() \%>\% select(id)
}
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
