% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.R
\name{ons_available_datasets}
\alias{ons_available_datasets}
\title{Available Datasets}
\usage{
ons_available_datasets()
}
\value{
list of available datasets
}
\description{
Available Datasets
}
\examples{
\dontrun{
ons_available_datasets()
}
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
