% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.R
\name{ons_available_versions}
\alias{ons_available_versions}
\title{Available Versions}
\usage{
ons_available_versions(id, edition)
}
\arguments{
\item{id}{dataset identifier. Valid values from \code{ons_available_datasets(...)}}

\item{edition}{edition identifier. Valid values from \code{ons_available_editions(...)}}
}
\value{
a list of version identifiers
}
\description{
Available Versions
}
\author{
neale
}
