% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.R
\name{ons_dataset_by_id}
\alias{ons_dataset_by_id}
\title{Dataset By Id}
\usage{
ons_dataset_by_id(df, id, edition, version)
}
\arguments{
\item{df}{dataframe describing the dataset}

\item{id}{the identifier of the dataset. Valid values from \code{ons_available_datasets()}}

\item{edition}{the edition of the dataset (if empty, select latest). Valid values from \code{ons_available_editions(...)}}

\item{version}{the version of the dataset (if empty, select latest). Valid values from \code{ons_available_available(...)}}
}
\value{
a dataframe describing the dataset.
}
\description{
Retrieve the metadata for the given dataset.
}
\details{
Makes calls to the ONS API and retrieves the metadata for the
datasets. The dataset selection can be refined via the edition and
version parameters
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
