% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.R
\name{ons_download}
\alias{ons_download}
\title{Download}
\usage{
ons_download(df, filebase, format = "csv")
}
\arguments{
\item{df}{dataframe}

\item{filebase}{base of the filename to which the data should be downloaded}

\item{format}{a valid format for the download}
}
\description{
\code{ons_download} retrieves the data described by the given df
}
