% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{thf_pipeline_defaults}
\alias{thf_pipeline_defaults}
\title{Create the THF defaults}
\usage{
thf_pipeline_defaults(download_root = "")
}
\arguments{
\item{download_root}{Root of directory hierarchy.}
}
\value{
an augmented metadata
}
\description{
Create the THF defaults
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
