% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{thf_read_file}
\alias{thf_read_file}
\title{Read the file described by the df}
\usage{
thf_read_file(df)
}
\arguments{
\item{df}{dataframe describing the downloaded file.}
}
\value{
a df incorporating the data. The actually data can then be
extracted with \code{\link{thf_data}}
}
\description{
Read the file described by the df
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
