% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{write_rds}
\alias{write_rds}
\title{write the data as a RDS.}
\usage{
write_rds(data, thf, create_directory)
}
\arguments{
\item{data}{The actual data}

\item{thf}{metadata dataframe created by the pipeline}

\item{create_directory}{boolean indicating whether to
(recursively) create the directory hierarchy.}
}
\value{
boolean indicating success
}
\description{
write the data as a RDS.
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
